support: move rename tools pkg to support
