+++ actions/shell/action.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/go-gilbert/gilbert-sdk"
-	"github.com/go-gilbert/gilbert/tools/shell"
+	"github.com/go-gilbert/gilbert/support/shell"
 )
 
 // Action represents sdk.ActionHandler
