+++ plugins/sources/http/handler.go
@@ -12,8 +12,8 @@ import (
 	"github.com/go-gilbert/gilbert/log"
 	"github.com/go-gilbert/gilbert/plugins/support"
 	"github.com/go-gilbert/gilbert/storage"
-	"github.com/go-gilbert/gilbert/tools/fs"
-	"github.com/go-gilbert/gilbert/tools/web"
+	"github.com/go-gilbert/gilbert/support/fs"
+	"github.com/go-gilbert/gilbert/support/web"
 )
 
 const (
