+++ support/shell/env.go
