+++ support/shell/shell.go
