+++ support/strings.go
@@ -1,4 +1,4 @@
-package tools
+package support
 
 import "strings"
